initial removal of attempt/candidate/etc.
